refactor: リファクタリングの第１歩
