+++ dev/20190704/k2works/api/rental/domain/movie.rb
@@ -1,0 +1,14 @@
+module Domain
+  class Movie
+    REGULAR = 0
+    NEW_RELEASE = 1
+    CHILDREN = 2
+
+    attr_reader :title
+    attr_reader :price_code
+
+    def initialize(title, price_code)
+      @title, @price_code = title, price_code
+    end
+  end
+end
