+++ dev/20190704/k2works/api/rental/domain/rental.rb
@@ -1,0 +1,8 @@
+module Domain
+  class Rental
+    attr_reader :movie, :days_rented
+    def initialize(movie, days_rented)
+      @movie, @days_rented = movie, days_rented
+    end
+  end
+end
