--- dev/20190704/k2works/api/rental/rental.rb
@@ -1,6 +1,0 @@
-class Rental
-  attr_reader :movie, :days_rented
-  def initialize(movie, days_rented)
-    @movie, @days_rented = movie, days_rented
-  end
-end
